+++ volunteer/models/__init__.py
@@ -2,6 +2,7 @@
 #
 # SPDX-License-Identifier: AGPL-3.0-or-later
 
+from . import volunteer_shift_mixin
 from . import volunteer_shift
 from . import volunteer_shift_category
 from . import volunteer_shift_type
