+++ volunteer/models/volunteer_shift.py
@@ -4,25 +4,19 @@
 
 from odoo import api, fields, models
 from odoo.exceptions import AccessError, ValidationError
-from odoo.tools import format_datetime
 from odoo.tools.translate import _
-
-from odoo.addons.base.models.res_partner import _tz_get
 
 
 class VolunteerShift(models.Model):
     _name = "volunteer.shift"
     _description = "Shift"
-    _inherit = ["mail.thread", "mail.activity.mixin"]
+    _inherit = ["volunteer.shift.mixin", "mail.thread", "mail.activity.mixin"]
 
     # General fields
 
-    name = fields.Char(required=True, tracking=True)
-    max_volunteer_nb = fields.Integer(
-        string="Max Volunteer", required=True, tracking=True
+    remaining_slots = fields.Integer(
+        compute="_compute_remaining_slots", store=True, tracking=True
     )
-    remaining_slots = fields.Integer(compute="_compute_remaining_slots", tracking=True)
-    is_one_day = fields.Boolean(compute="_compute_is_one_day")
 
     # Stage fields
 
@@ -39,55 +33,17 @@ class VolunteerShift(models.Model):
     )
     state = fields.Selection(related="stage_id.state", store=True)
 
-    # Date fields
-
-    tz = fields.Selection(
-        selection=_tz_get,
-        string="Timezone",
-        default=lambda self: self.env.user.tz or "UTC",
-        required=True,
-        tracking=True,
-    )
-    start_time = fields.Datetime(
-        default=fields.Datetime.now(), required=True, tracking=True
-    )
-    start_time_located = fields.Char(compute="_compute_start_time_located")
-    end_time = fields.Datetime(
-        default=fields.Datetime.now(), required=True, tracking=True
-    )
-    end_time_located = fields.Char(compute="_compute_end_time_located")
-
-    # Classification fields
-
-    type_id = fields.Many2one(
-        comodel_name="volunteer.shift.type", string="Type", required=True, tracking=True
-    )
-    category_id = fields.Many2one(
-        comodel_name="volunteer.shift.category", string="Category", tracking=True
-    )
-    tag_ids = fields.Many2many(
-        comodel_name="volunteer.shift.tag", string="Tags", tracking=True
-    )
-
     # Relational fields
 
-    company_id = fields.Many2one(
-        comodel_name="res.company",
-        string="Company",
-        default=lambda self: self.env.user.company_id,
-        required=True,
-        tracking=True,
-    )
     volunteer_participation_ids = fields.One2many(
         comodel_name="volunteer.shift.participation",
         inverse_name="shift_id",
         string="Participation",
         tracking=True,
     )
-    coordinator_id = fields.Many2one(
-        comodel_name="res.partner",
-        domain="[('is_company', '=', False)]",
-        string="Coordinator",
+    generator_id = fields.Many2one(
+        comodel_name="volunteer.shift.recurrent.generator",
+        string="Shift Generator",
         tracking=True,
     )
     volunteer_ids = fields.Many2many(
@@ -97,60 +53,25 @@ class VolunteerShift(models.Model):
         store=True,
     )
 
-    # SQL constraints
-
     _sql_constraints = [
         (
-            "max_volunteer_nb_is_positive",
+            "shift_max_vol_nb_is_pos",
             "check (max_volunteer_nb > 0)",
-            "The maximum of volunteers per shift cannot be null or negative.",
+            "The maximum of volunteers cannot be null or negative.",
         ),
     ]
 
     # Compute methods
 
-    @api.depends("volunteer_participation_ids")
+    @api.depends(
+        "volunteer_participation_ids", "volunteer_participation_ids.registration_state"
+    )
     def _compute_remaining_slots(self):
         for shift in self:
             nb_confirmed_participation = shift.get_booking_status()[
                 "nb_confirmed_participation"
             ]
             shift.remaining_slots = shift.max_volunteer_nb - nb_confirmed_participation
-
-    @api.depends("start_time", "end_time", "tz")
-    def _compute_is_one_day(self):
-        for shift in self:
-            shift = shift._set_tz_context()
-            if shift.start_time and shift.end_time:
-                start_date = fields.Datetime.context_timestamp(
-                    shift, shift.start_time
-                ).date()
-                end_date = fields.Datetime.context_timestamp(
-                    shift, shift.end_time
-                ).date()
-                shift.is_one_day = start_date == end_date
-            else:
-                shift.is_one_day = False
-
-    @api.depends("tz", "start_time")
-    def _compute_start_time_located(self):
-        for shift in self:
-            if shift.start_time:
-                shift.start_time_located = format_datetime(
-                    self.env, shift.start_time, shift.tz, dt_format="medium"
-                )
-            else:
-                shift.start_time_located = False
-
-    @api.depends("tz", "end_time")
-    def _compute_end_time_located(self):
-        for shift in self:
-            if shift.end_time:
-                shift.end_time_located = format_datetime(
-                    self.env, shift.end_time, shift.tz, dt_format="medium"
-                )
-            else:
-                shift.end_time_located = False
 
     @api.depends("volunteer_participation_ids")
     def _compute_volunteer_ids(self):
@@ -201,18 +122,19 @@ class VolunteerShift(models.Model):
     # Override methods
 
     def write(self, vals):
-        state_requested = vals.get("state")
+        old_states = {shift.id: shift.state for shift in self}
         # Restrict stage change to admins only
         if (
             "stage_id" in vals
             and not self.env.context.get("install_mode")
             and not self.env.user.has_group("volunteer.volunteer_group_admin")
         ):
-            raise AccessError(_("Only admins can change the stage of a shift"))
+            raise AccessError(_("Only admins can change the state of a shift"))
         res = super().write(vals)
         for shift in self:
+            previous_state = old_states[shift.id]
             # Auto-cancel confirmed participation if the shift is canceled
-            if state_requested != "canceled" and shift.state == "canceled":
+            if previous_state != "canceled" and shift.state == "canceled":
                 confirmed_participation = shift.get_booking_status()[
                     "confirmed_participation"
                 ]
@@ -224,11 +146,6 @@ class VolunteerShift(models.Model):
     @api.model
     def _group_expand_stage_id(self, stages, domain, order):
         return stages.search([], order=order)
-
-    def _set_tz_context(self):
-        """Set the timezone context for the shift."""
-        self.ensure_one()
-        return self.with_context(tz=self.tz)
 
     def get_booking_status(self):
         """Get the shift booking status by returning a dictionary of :
